refac: useCoupon - state, storage 로직 분리
