+++ src/basic/hooks/useCoupons.ts
@@ -1,13 +1,26 @@
-// TODO: 쿠폰 관리 Hook
-// 힌트:
-// 1. 쿠폰 목록 상태 관리 (localStorage 연동 고려)
-// 2. 쿠폰 추가/삭제
-//
-// 반환할 값:
-// - coupons: 쿠폰 배열
-// - addCoupon: 새 쿠폰 추가
-// - removeCoupon: 쿠폰 삭제
+import { useState, useEffect } from "react";
+import { Coupon } from "../types";
+import { INITIAL_COUPONS } from "../constants";
 
+// 1단계: 기본 상태 관리만
 export function useCoupons() {
-  // TODO: 구현
+  // 쿠폰 목록 상태 (localStorage에서 복원)
+  const [coupons, setCoupons] = useState<Coupon[]>(() => {
+    const saved = localStorage.getItem("coupons");
+    if (saved) {
+      try {
+        return JSON.parse(saved);
+      } catch {
+        return INITIAL_COUPONS;
+      }
+    }
+    return INITIAL_COUPONS;
+  });
+
+  // localStorage 동기화
+  useEffect(() => {
+    localStorage.setItem("coupons", JSON.stringify(coupons));
+  }, [coupons]);
+
+  return { coupons, setCoupons };
 }
